moves label files from feature to components folder
